Day 1's Produce!
Okay here we are, end of day 1! Here I've thrown together a basic
overworld simulation. You can roam around this horribly tiled castle
(level design is something for me to improve on) and speak to 2 NPCs.
The dots on the bottom of your sprite are there to show which direction
your character is facing. The blue square floating around your character
I suppose is your speech range, it's a little object I threw together
that allows the game to check if you are standing close enough to an NPC
to be able to talk to them, of course this will be made invisible later.
To initiate a conversation with an NPC, stand close enough so that the
blue square is touching them and press Z. And uh.. yeah that's all there
is really. I'd better sleep now.